feat: Complete restructure of the code base
