+++ src/OnBeat.h
@@ -1,0 +1,14 @@
+#pragma once
+
+//Include all headers in onbeat
+
+#include "OnBeat/App/App.h"
+#include "OnBeat/App/Layer/MusicLayer.h"
+
+#include "OnBeat/Config/Config.h"
+#include "OnBeat/Config/Skin.h"
+
+#include "OnBeat/Ui/MainMenu/MainMenu.h"
+#include "OnBeat/Ui/PauseMenu/PauseMenu.h"
+
+#include "OnBeat/Util/Util.h"
