+++ src/OnBeat/Ui/PauseMenu/PauseMenu.cpp
@@ -1,0 +1,14 @@
+#include <OnBeat/Ui/PauseMenu/PauseMenu.h>
+
+namespace OnBeat
+{
+	PauseMenu::PauseMenu(const std::string& document, App* MainApp, std::string layerName)
+		: Menu(document, MainApp, layerName)
+	{
+	}
+
+	PauseMenu::~PauseMenu()
+	{
+
+	}
+}
