+++ src/OnBeat/Util/Util.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "OnSetDetection/OnSetDetection.h"
+
+#include "AudioPlayer/AudioPlayer.h"
+
+#include "AppUtil/AppUtil.h"
+
+#include "Template/GLTextureSurface.h"
+#include "Template/Layer.h"
+#include "Template/Menu.h"
